neribot/requirements.txt
aiocache==0.12.3
aiofiles==23.2.1
aiosqlite==0.17.0
anyio==4.7.0
apscheduler==3.11.0
arclet-alconna-tools==0.7.10
arclet-alconna==1.8.35
arrow==1.3.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
beautifulsoup4==4.12.3
bilireq==0.2.3.post0
binaryornot==0.4.4
cachetools==5.5.0
cashews==7.4.0
cattrs==23.2.3
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cn2an==0.5.22
colorama==0.4.6
cookiecutter==2.6.0
cryptography==44.0.0
dateparser==1.2.0
distlib==0.3.9
ecdsa==0.19.0
emoji==2.14.0
exceptiongroup==1.2.2
fastapi==0.115.6
feedparser==6.0.11
filelock==3.16.1
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
httpcore==0.16.3
httptools==0.6.4
httpx==0.23.3
idna==3.10
imagehash==4.3.1
importlib-metadata==8.5.0
iso8601==1.1.0
jinja2==3.1.4
loguru==0.7.3
lxml==5.3.0
markdown-it-py==3.0.0
markdown==3.7
markupsafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
nb-cli==1.4.2
nepattern==0.7.7
nonebot-adapter-discord==0.1.8
nonebot-adapter-dodo==0.1.4
nonebot-adapter-kaiheila==0.3.4
nonebot-adapter-onebot==2.4.6
nonebot-plugin-alconna==0.54.1
nonebot-plugin-apscheduler==0.3.0
nonebot-plugin-htmlrender==0.3.5
nonebot-plugin-session==0.2.3
nonebot-plugin-uninfo==0.4.1
nonebot-plugin-userinfo==0.1.3
nonebot-plugin-waiter==0.8.0
nonebot2==2.4.0
nonebot2[fastapi]==2.4.0
noneprompt==0.1.9
numpy==2.2.0
platformdirs==4.3.6
playwright==1.49.1
proces==0.1.7
prompt-toolkit==3.0.48
propcache==0.2.1
protobuf==4.25.5
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.18
pyee==12.0.0
pyfiglet==1.0.2
pygments==2.18.0
pygtrie==2.5.0
pymdown-extensions==10.12
pypika-tortoise==0.1.6
pypinyin==0.51.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-markdown-math==0.8
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.2
pywavelets==1.8.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3986[idna2008]==1.5.0
rich==13.9.4
rsa==4.9
ruamel-yaml-clib==0.2.12
ruamel-yaml==0.18.6
scipy==1.14.1
setuptools==75.6.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
strenum==0.4.15
tarina==0.6.8
text-unidecode==1.3
tomli==2.2.1
tomlkit==0.13.2
tortoise-orm[asyncpg]==0.20.0
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
uvicorn[standard]==0.32.1
uvloop==0.21.0
virtualenv==20.28.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.1
win32-setctime==1.2.0
yarl==1.18.3
zipp==3.21.0
jieba==0.42.1
matplotlib==3.10.0
nonebot==1.9.1
nonebot_plugin_apscheduler==0.5.0
nonebot_plugin_session==0.3.2
typing_extensions==4.12.2
wordcloud==1.9.4
